Update purchased price when an asset is bought

+++ blockchain/src/main/java/com/hackathon/blockchain/model/Asset.java
@@ -24,7 +24,7 @@ public class Asset {
     private String symbol;
     private Double quantity;
 
-    private Integer purchasedPrice;
+    private Double purchasedPrice;
 
     @ManyToOne
     private Wallet wallet;
